fix: alarmTerm Long -> int

+++ src/main/java/com/seulseul/seulseul/dto/alarm/AlarmDto.java
@@ -10,7 +10,7 @@ public class AlarmDto {
     private Long id;
     private boolean alarmEnabled;
     private Long alarmTime;
-    private Long alarmTerm;
+    private int alarmTerm;
 
     public AlarmDto(Alarm alarm) {
         this.id = alarm.getId();
